backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pymysql==1.1.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4.1
# bcrypt 4.x quebra o passlib 1.7.4 silenciosamente (verify_password sempre retorna False)
# Manter na 3.2.2 até atualizar para argon2 ou superior
bcrypt==3.2.2
python-multipart==0.0.20
email-validator==2.2.0
aiofiles==24.1.0
